moved ListViewControllerTests to ListViewControllerUITests to use snapshot testing instead
